最近联系人数量增加

+++ src/main/java/cn/shu/wechat/entity/Contacts.java
@@ -67,7 +67,7 @@ public class Contacts {
     private Double unifriend;
 
     @Excel(name="contactflag")
-    private Double contactflag;
+    private int contactflag;
 
     //@ExcelCollection(name="群成员")
     private List<Contacts> memberlist;
